owmweather - Move pulling code to lib
