+++ apps/owmweather/metadata.json
@@ -13,6 +13,7 @@
   "storage": [
     {"name":"owmweather.default.json","url":"default.json"},
     {"name":"owmweather.boot.js","url":"boot.js"},
+    {"name":"owmweather","url":"lib.js"},
     {"name":"owmweather.settings.js","url":"settings.js"}
   ]
 }
